fix: change char and i8 to u8
The V language has gone through some changes after this code has been
written, and one of them is the change from char data type to u8.
Therefore, there was a need to update the code to reflect this. char was
renamed to u8, and the previous return type of i8 is now u8.